// core/ledger-client/src/generated-clients/openapi-3.4.0-SNAPSHOT.ts
export interface paths {
    '/v2/commands/submit-and-wait': {
        /** @description Submit a batch of commands and wait for the completion details */
        post: operations['postV2CommandsSubmit-and-wait']
    }
// ...
    '/v2/parties': {
        /** @description List all known parties. */
        get: operations['getV2Parties']
        /** @description Allocate a new party to the participant node */
        post: operations['postV2Parties']
    }
    '/v2/parties/participant-id': {
        /** @description Get participant id */
        get: operations['getV2PartiesParticipant-id']
    }
// ...
    '/v2/users/{user-id}/rights': {
        /** @description List user rights. */
        get: operations['getV2UsersUser-idRights']
        /** @description Grant user rights. */
        post: operations['postV2UsersUser-idRights']
        /** @description Revoke user rights. */
        patch: operations['patchV2UsersUser-idRights']
    }
// ...
    '/v2/interactive-submission/prepare': {
        /** @description Prepare commands for signing */
        post: operations['postV2Interactive-submissionPrepare']
    }
// ...
}

export type webhooks = Record<string, never>

export interface components {
// ...
    responses: never
    parameters: never
    requestBodies: never
    headers: never
    pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
    /** @description Submit a batch of commands and wait for the completion details */
    'postV2CommandsSubmit-and-wait': {
        requestBody: {
            content: {
                'application/json': components['schemas']['JsCommands']
            }
        }
        responses: {
            200: {
                content: {
                    'application/json': components['schemas']['SubmitAndWaitResponse']
                }
            }
            /** @description Invalid value for: body, Invalid value for: headers */
            400: {
                content: {
                    'text/plain': string
                }
            }
            default: {
                content: {
                    'application/json': components['schemas']['JsCantonError']
                }
            }
        }
    }
// ...
}